requirements.txt
pandas
tensorflow-cpu